+++ docs/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project (aims to) adhere to [Semantic Versioning](https://semver.org/sp
 ### Fixed
 - themed busy cursor on native Wayland
 - ignore cue-referenced disc image files
+- select newly imported GOG library item
 
 ---